split test cases in 2 modules so that eduid-am can import form the base one w/o pulling flask stuff
